+++ harness/determined/pytorch/lightning/_adapter.py
@@ -21,8 +21,8 @@ from determined.pytorch import (
     PyTorchCallback,
     PyTorchTrial,
     PyTorchTrialContext,
-    tensorboard_writer,
 )
+from determined.pytorch.tensorboard import TorchWriter
 from determined.util import filter_duplicates, has_param
 
 TorchData = Union[Dict[str, torch.Tensor], Sequence[torch.Tensor], torch.Tensor]
@@ -78,7 +78,7 @@ def check_compatibility(lm: pl.LightningModule) -> None:
 
 
 def override_unsupported_nud(lm: pl.LightningModule, context: PyTorchTrialContext) -> None:
-    writer = tensorboard_writer.TorchWriter()
+    writer = TorchWriter()
 
     def lm_print(*args: Any, **kwargs: Any) -> None:
         if context.distributed.get_rank() == 0:
