+++ harness/determined/tensorboard/base.py
@@ -132,9 +132,9 @@ def get_metric_writer() -> tensorboard.BatchMetricWriter:
 
     except ModuleNotFoundError:
         logging.warning("TensorFlow writer not found")
-        from determined.pytorch import tensorboard_writer
+        from determined.pytorch.tensorboard import TorchWriter
 
-        writer = tensorboard_writer.TorchWriter()
+        writer = TorchWriter()
 
     return tensorboard.BatchMetricWriter(writer)
